// src/api/schema.d.ts
export interface components {
    schemas: {
        id_name: {
            id: number;
            name: string;
        };
        "issue_status.summary": {
            id: number;
            name: string;
            is_closed: boolean;
        };
        custom_field_value: {
            id: number;
            name: string;
            multiple?: boolean;
            value: (string | null) | string[];
        };
        attachment: {
            id: number;
            filename: string;
            filesize: number;
            content_type: string | null;
            description: string | null;
            content_url: string;
            thumbnail_url?: string;
            author?: components["schemas"]["id_name"];
            /** Format: date-time */
            created_on: string;
        };
// ...
    };
// ...
}